third_party/WebKit/LayoutTests/animations/usecounters/use-count-stacked-neutral-keyframe.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#target {
    animation-name: a, b;
    animation-duration: 1e10s;
}

@keyframes a {
    from { left: 1px; }
    to { left: 1px; }
}

@keyframes b {
    from { left: 1px; }
    to { left: 1px; }
}
</style>
<div id="target"></div>
<script>
var CSSAnimationsStackedNeutralKeyframe = 937; // From UseCounter.h
test(function() {
  getComputedStyle(target).left; // Force style recalc;
  assert_false(internals.isUseCounted(document, CSSAnimationsStackedNeutralKeyframe));

  document.styleSheets[0].rules[2].deleteRule('from');
  getComputedStyle(target).left; // Force style recalc;
  assert_true(internals.isUseCounted(document, CSSAnimationsStackedNeutralKeyframe));
});
</script>
